Remove lazy_static dependency and usage of locks
This removes the usage of two `RwLock<HashMap<_, _>>`s.

The issue is that while one WebSocket client was reading data and
parsing it, that client would hold a write lock to the hashmap. This
blocked other clients from accessing the hashmap to retrieve the
client's entry. This means that if many WebSocket clients are active (on
the scale of 5+, especially 15+), they would become unusable.

Clients now maintain owned packet and reader states.